refactor: 유틸 예외 처리 통일
