+++ src/main/java/kr/co/fastcampus/fastcatch/domain/cart/service/CartService.java
@@ -1,6 +1,7 @@
 package kr.co.fastcampus.fastcatch.domain.cart.service;
 
 import java.util.Optional;
+import kr.co.fastcampus.fastcatch.common.exception.NoCartItemException;
 import kr.co.fastcampus.fastcatch.common.utility.AvailableOrderUtil;
 import kr.co.fastcampus.fastcatch.domain.accommodation.entity.Room;
 import kr.co.fastcampus.fastcatch.domain.accommodation.service.AccommodationService;
@@ -8,7 +9,6 @@ import kr.co.fastcampus.fastcatch.domain.cart.dto.request.CartItemRequest;
 import kr.co.fastcampus.fastcatch.domain.cart.dto.response.CartResponse;
 import kr.co.fastcampus.fastcatch.domain.cart.entity.Cart;
 import kr.co.fastcampus.fastcatch.domain.cart.entity.CartItem;
-import kr.co.fastcampus.fastcatch.common.exception.NoCartItemException;
 import kr.co.fastcampus.fastcatch.domain.cart.repository.CartItemRepository;
 import kr.co.fastcampus.fastcatch.domain.cart.repository.CartRepository;
 import kr.co.fastcampus.fastcatch.domain.member.service.MemberService;
